made session save only if more than 3 questions have been done
