+++ app/(tabs)/inputSessionNotes.tsx
@@ -7,6 +7,7 @@ import { router, useLocalSearchParams } from "expo-router";
 import { useEffect, useState } from "react";
 import { StyleSheet, TextInput, TouchableOpacity, View } from "react-native";
 import uuid from "react-native-uuid";
+import { handleSubmission} from "@/lib/saveSession";
 
 export interface SessionDetails {
   notes: string;
@@ -27,7 +28,7 @@ export default function InputSessionNotes() {
     const newUUID = uuid.v4();
     console.log("Generated UUID:", newUUID);
     return newUUID;
-  };
+  };  
 
   useEffect(() => {
     const currentDate = new Date();
@@ -54,86 +55,6 @@ export default function InputSessionNotes() {
       }
     }
   }, [sessionData]);
-
-  async function createTestSession(sessionUUID: string, childId: string) {
-    try {
-      const { data, error } = await supabase
-        .from("test_session")
-        .insert({ id: sessionUUID, child_id: childId })
-        .select()
-        .single();
-
-      if (error) {
-        console.log("Error fetching test session:", error.message);
-        return null;
-      }
-      console.log("Test session created:", data);
-      return data;
-    } catch (error) {
-      console.error("Error creating test session:", (error as Error).message);
-      return null;
-    }
-  }
-
-  async function insertAttempts(sessionUUID: string) {
-    const trials = parsedAttempts.map((trial) => ({
-      prompt: trial.prompt,
-      response: trial.response,
-      test_session_id: sessionUUID,
-    }));
-
-    const { data, error } = await supabase.from("test_trial").insert(trials);
-
-    if (error) {
-      console.log("Error fetching test trial:", error.message);
-      return null;
-    }
-    console.log("Successfully inserted trials");
-  }
-
-  async function insertSettingsData(sessionUUID: string) {
-    try {
-      const { data, error } = await supabase
-        .from("test_session_settings")
-        .insert({
-          id: sessionUUID,
-          set_size: numCards,
-          sound_enabled: soundEnabled,
-          max_thresh_level: maximumThresholdLevel,
-          session_notes: sessionNotes,
-        })
-        .select()
-        .single();
-      if (error) {
-        console.log("Error creating settings data:", error.message);
-      }
-      console.log("Test session created:", data);
-      return data;
-    } catch (error) {
-      console.error("Error creating test session:", (error as Error).message);
-      return null;
-    }
-  }
-
-  async function handleSubmission() {
-    const userUUID = userData.CURRENT_ID;
-    const sessionUUID = generateUUID();
-
-    try {
-      const newSession = await createTestSession(sessionUUID, userUUID);
-
-      if (!newSession) {
-        throw new Error("Failed to create test session");
-      }
-
-      await insertAttempts(sessionUUID);
-      await insertSettingsData(sessionUUID);
-
-      console.log("Test session created and attempts inserted successfully");
-    } catch (error) {
-      console.error("Error in submission process:", error);
-    }
-  }
 
   return (
     <View style={styles.wrapper}>
@@ -173,7 +94,7 @@ export default function InputSessionNotes() {
           <TouchableOpacity
             style={styles.button}
             onPress={async () => {
-              await handleSubmission();
+              await handleSubmission(parsedAttempts,numCards,soundEnabled,maximumThresholdLevel,sessionNotes);
               router.push({
                 pathname: "/(tabs)/sessionResults",
                 params: {
